@@ -10,28 +10,14 @@ using Machine.Specifications.Runner.ReSharper.Tasks;
 
 namespace Machine.Specifications.Runner.ReSharper.Adapters.Discovery;
 
-public class Discoverer
+public class Discoverer(
+    TestRequest request,
+    IMspecController controller,
+    ITestDiscoverySink sink,
+    RemoteTaskDepot depot,
+    CancellationToken token)
 {
-    private readonly TestRequest request;
-
-    private readonly IMspecController controller;
-
-    private readonly ITestDiscoverySink sink;
-
-    private readonly RemoteTaskDepot depot;
-
-    private readonly CancellationToken token;
-
     private readonly ILogger logger = Logger.GetLogger<Discoverer>();
-
-    public Discoverer(TestRequest request, IMspecController controller, ITestDiscoverySink sink, RemoteTaskDepot depot, CancellationToken token)
-    {
-        this.request = request;
-        this.controller = controller;
-        this.sink = sink;
-        this.depot = depot;
-        this.token = token;
-    }
 
     public void Discover()
     {
