@@ -2,18 +2,8 @@
 
 namespace Machine.Specifications.Runner.ReSharper.Adapters.Listeners;
 
-internal class AdapterListener : ISpecificationRunListener
+internal class AdapterListener(IRunListener listener, string assemblyPath) : ISpecificationRunListener
 {
-    private readonly IRunListener listener;
-
-    private readonly string assemblyPath;
-
-    public AdapterListener(IRunListener listener, string assemblyPath)
-    {
-        this.listener = listener;
-        this.assemblyPath = assemblyPath;
-    }
-
     public void OnAssemblyStart(AssemblyInfo assemblyInfo)
     {
         listener.OnAssemblyStart(new TestAssemblyInfo(assemblyPath));
