@@ -1,17 +1,10 @@
 ﻿namespace Machine.Specifications.Runner.ReSharper.Adapters.Listeners;
 
-public class TestSpecificationInfo
+public class TestSpecificationInfo(string containingType, string fieldName, string capturedOutput)
 {
-    public TestSpecificationInfo(string containingType, string fieldName, string capturedOutput)
-    {
-        ContainingType = containingType;
-        FieldName = fieldName;
-        CapturedOutput = capturedOutput;
-    }
+    public string ContainingType { get; } = containingType;
 
-    public string ContainingType { get; }
+    public string FieldName { get; } = fieldName;
 
-    public string FieldName { get; }
-
-    public string CapturedOutput { get; }
+    public string CapturedOutput { get; } = capturedOutput;
 }
