@@ -4,21 +4,18 @@ using JetBrains.ReSharper.TestRunner.Abstractions.Objects;
 namespace Machine.Specifications.Runner.ReSharper.Tasks;
 
 [Serializable]
-public abstract class MspecRemoteTask : RemoteTask
+public abstract class MspecRemoteTask(
+    string testId,
+    string? ignoreReason,
+    bool runAllChildren = true,
+    bool runExplicitly = false)
+    : RemoteTask
 {
-    protected MspecRemoteTask(string testId, string? ignoreReason, bool runAllChildren = true, bool runExplicitly = false)
-    {
-        TestId = testId;
-        IgnoreReason = ignoreReason;
-        RunAllChildren = runAllChildren;
-        RunExplicitly = runExplicitly;
-    }
+    public string TestId { get; } = testId;
 
-    public string TestId { get; }
+    public string? IgnoreReason { get; } = ignoreReason;
 
-    public string? IgnoreReason { get; }
+    public bool RunAllChildren { get; } = runAllChildren;
 
-    public bool RunAllChildren { get; }
-
-    public bool RunExplicitly { get; }
+    public bool RunExplicitly { get; } = runExplicitly;
 }
