@@ -1,5 +1,4 @@
 ﻿using System.Linq;
-using JetBrains.ReSharper.TestRunner.Abstractions.Objects;
 using Machine.Specifications.Runner.ReSharper.Adapters;
 using Machine.Specifications.Runner.ReSharper.Tests.Fixtures;
 using NUnit.Framework;
@@ -12,75 +11,69 @@ public class RemoteTaskDepotTests
     [Test]
     public void CanGetContextByElement()
     {
-        var depot = new RemoteTaskDepot(new RemoteTask[]
-        {
+        var depot = new RemoteTaskDepot([
             RemoteTaskFixtures.Context
-        });
+        ]);
 
-        Assert.NotNull(depot[ElementFixtures.Context]);
+        Assert.That(depot[ElementFixtures.Context], Is.Not.Null);
     }
 
     [Test]
     public void CanGetSpecificationByElement()
     {
-        var depot = new RemoteTaskDepot(new RemoteTask[]
-        {
+        var depot = new RemoteTaskDepot([
             RemoteTaskFixtures.Context,
             RemoteTaskFixtures.Specification1
-        });
+        ]);
 
-        Assert.NotNull(depot[ElementFixtures.Specification1]);
+        Assert.That(depot[ElementFixtures.Specification1], Is.Not.Null);
     }
 
     [Test]
     public void CanGetBehaviorByElement()
     {
-        var depot = new RemoteTaskDepot(new RemoteTask[]
-        {
+        var depot = new RemoteTaskDepot([
             RemoteTaskFixtures.Context,
             RemoteTaskFixtures.Behavior1
-        });
+        ]);
 
-        Assert.NotNull(depot[ElementFixtures.Behavior1]);
+        Assert.That(depot[ElementFixtures.Behavior1], Is.Not.Null);
     }
 
     [Test]
     public void CanGetBehaviorSpecificationByElement()
     {
-        var depot = new RemoteTaskDepot(new RemoteTask[]
-        {
+        var depot = new RemoteTaskDepot([
             RemoteTaskFixtures.Context,
             RemoteTaskFixtures.Behavior1,
             RemoteTaskFixtures.Behavior1Specification1
-        });
+        ]);
 
-        Assert.NotNull(depot[ElementFixtures.Behavior1Specification1]);
+        Assert.That(depot[ElementFixtures.Behavior1Specification1], Is.Not.Null);
     }
 
     [Test]
     public void CanGetBehaviorWhenThereIsSpecificationWithSameName()
     {
-        var depot = new RemoteTaskDepot(new RemoteTask[]
-        {
+        var depot = new RemoteTaskDepot([
             RemoteTaskFixtures.Context,
             RemoteTaskFixtures.Behavior1,
             RemoteTaskFixtures.Specification1,
             RemoteTaskFixtures.Behavior1Specification1
-        });
+        ]);
 
-        Assert.NotNull(depot[ElementFixtures.Specification1]);
+        Assert.That(depot[ElementFixtures.Specification1], Is.Not.Null);
     }
 
     [Test]
     public void BoundElementsAreRunnable()
     {
-        var depot = new RemoteTaskDepot(new RemoteTask[]
-        {
+        var depot = new RemoteTaskDepot([
             RemoteTaskFixtures.Context,
             RemoteTaskFixtures.Behavior1,
             RemoteTaskFixtures.Specification1,
             RemoteTaskFixtures.Behavior1Specification1
-        });
+        ]);
 
         depot.Bind(ElementFixtures.Context, RemoteTaskFixtures.Context);
         depot.Bind(ElementFixtures.Behavior1, RemoteTaskFixtures.Behavior1);
@@ -89,7 +82,7 @@ public class RemoteTaskDepotTests
 
         var selected = depot.GetTestsToRun().ToArray();
 
-        CollectionAssert.Contains(selected, ElementFixtures.Specification1);
-        CollectionAssert.Contains(selected, ElementFixtures.Behavior1Specification1);
+        Assert.That(selected, Has.Member(ElementFixtures.Specification1));
+        Assert.That(selected, Has.Member(ElementFixtures.Behavior1Specification1));
     }
 }
