@@ -1,4 +1,5 @@
 ﻿using System;
+using JetBrains.Application.Parts;
 using JetBrains.ProjectModel;
 using JetBrains.ReSharper.UnitTestFramework.Execution;
 using JetBrains.Util;
@@ -6,15 +7,10 @@ using Machine.Specifications.Runner.ReSharper.Runner;
 
 namespace Machine.Specifications.Runner.ReSharper;
 
-[SolutionComponent]
-public class MspecServiceProvider
+[SolutionComponent(Instantiation.ContainerAsyncPrimaryThread)]
+public class MspecServiceProvider(ISolution solution)
 {
-    private readonly Lazy<MspecTestRunnerRunStrategy> runner;
-
-    public MspecServiceProvider(ISolution solution)
-    {
-        runner = Lazy.Of(solution.GetComponent<MspecTestRunnerRunStrategy>, true);
-    }
+    private readonly Lazy<MspecTestRunnerRunStrategy> runner = Lazy.Of(solution.GetComponent<MspecTestRunnerRunStrategy>, true);
 
     public IUnitTestRunStrategy GetRunStrategy()
     {
