@@ -1,4 +1,5 @@
-﻿using JetBrains.ProjectModel;
+﻿using JetBrains.Application.Parts;
+using JetBrains.ProjectModel;
 using JetBrains.ProjectModel.NuGet.Packaging;
 using JetBrains.ReSharper.UnitTestFramework.Execution.TestRunner;
 using JetBrains.ReSharper.UnitTestFramework.Exploration.Artifacts;
@@ -6,15 +7,10 @@ using JetBrains.Util.Collections;
 
 namespace Machine.Specifications.Runner.ReSharper;
 
-[SolutionComponent]
-public class MspecTestExplorerFromTestRunner : UnitTestExplorerFrom.TestRunner
-{
-    public MspecTestExplorerFromTestRunner(
-        MspecTestProvider provider,
-        ITestRunnerAgentManager agentManager,
-        MspecTestRunnerOrchestrator adapter,
-        NuGetInstalledPackageChecker installedPackageChecker)
-        : base(provider, agentManager, adapter, installedPackageChecker, 1.Minute())
-    {
-    }
-}
+[SolutionComponent(Instantiation.ContainerAsyncPrimaryThread)]
+public class MspecTestExplorerFromTestRunner(
+    MspecTestProvider provider,
+    ITestRunnerAgentManager agentManager,
+    MspecTestRunnerOrchestrator adapter,
+    NuGetInstalledPackageChecker installedPackageChecker)
+    : UnitTestExplorerFrom.TestRunner(provider, agentManager, adapter, installedPackageChecker, 1.Minute());
