@@ -1,15 +1,11 @@
-﻿using JetBrains.ProjectModel;
+﻿using JetBrains.Application.Parts;
+using JetBrains.ProjectModel;
 using JetBrains.ReSharper.UnitTestFramework.Execution.TestRunner;
 
 namespace Machine.Specifications.Runner.ReSharper.Runner;
 
-[SolutionComponent]
-public class AgentManagerHost : IAgentManagerHost
+[SolutionComponent(Instantiation.ContainerAsyncPrimaryThread)]
+public class AgentManagerHost(ITestRunnerAgentManager testRunnerAgentManager) : IAgentManagerHost
 {
-    public AgentManagerHost(ITestRunnerAgentManager testRunnerAgentManager)
-    {
-        AgentManager = testRunnerAgentManager;
-    }
-
-    public ITestRunnerAgentManager AgentManager { get; }
+    public ITestRunnerAgentManager AgentManager { get; } = testRunnerAgentManager;
 }
